geom: assert the two station-level spellings agree to the bound invariance admits
Review round 1, Important 2 and Minor 8.

Making harmonic() read secondHarmonic().Level was the right production change, but it turned the row
that had motivated it into a tautology: both sides of the level assertion became literally
torusStationAt(...).secondHarmonic().Level, and the docstring still claimed to demonstrate a
reproduction it no longer demonstrates.

The two spellings are not the same number. axisInvariantEntries admits |m11−m22| ≤ axisInvarianceTol·
scale before it calls a tensor invariant at all, and the spellings differ by exactly ρ²(m11−m22)/2, so
every root() of the invariant family may move by that much for a NEAR-invariant tensor — a fact the
old bit-for-bit assertion only ever exercised on five exactly-invariant fixtures. The row now computes
the closed form's own arithmetic and compares, with the bound DERIVED from that admission plus the two
spellings' own roundings in ulps. Verified it can fail: a 1e-10 relative perturbation of the closed
form is caught at 3.8e-09 against a 3.3e-11 bound.

Minor 8: the WHY comment on Level claimed the explicit float64() stopped a fusion, but the term ends
in a division and a division result is not a product, so no platform could fuse it there. The
conversion is dropped and the comment now states the real reason this spelling is the stable one — and
warns to keep the division last if the line is rewritten.

+++ kernel/geom/torus_quadric_harmonic2.go
@@ -135,12 +135,14 @@ func (st torusStation) secondHarmonic() torusSecondHarmonic {
 		Sin2: rr * st.m12,
 		Cos1: st.rho * st.x,
 		Sin1: st.rho * st.y,
-		// float64(...) rounds the term BEFORE it is added, which the Go spec guarantees an explicit
-		// conversion does. It is the one term of this station a SECOND caller also reads (harmonic,
-		// above), so it is a value two paths must agree on bit for bit; without the conversion a
-		// platform that fuses x*y+z is free to carry it at full precision and round once, and the two
-		// paths part by an ulp (CI run 34280554924 macos-latest).
-		Level: st.constant + float64(rr*(st.m11+st.m22)/2),
+		// The term added to the constant ends in a DIVISION, and a division result is not a product,
+		// so no platform may fuse it into this add — which is what makes this spelling of the level the
+		// stable one and the closed form's `constant + m11·ρ²` the one that parted by an ulp on arm64
+		// (CI run 34280554924 macos-latest). harmonic() reads this Level rather than respelling it for
+		// exactly that reason. Keep the division last if this line is ever rewritten; an explicit
+		// float64() round of the term would pin it too, and was dropped only because it cannot fire
+		// here and a conversion that documents a mechanism it does not use reads as a live guard.
+		Level: st.constant + rr*(st.m11+st.m22)/2,
 	}
 }
